+++ R/RcppExports.R
@@ -192,8 +192,8 @@ vaeInnerLik <- function(etaMat, cores, grad = FALSE, preds = FALSE) {
     .Call(`_nlmixr2est_vaeInnerLik`, etaMat, cores, grad, preds)
 }
 
-vaeOuterSolve_ <- function(thVals, ebes, cols, cores) {
-    .Call(`_nlmixr2est_vaeOuterSolve_`, thVals, ebes, cols, cores)
+vaeOuterSolve_ <- function(thVals, ebes, cols, cores, tol) {
+    .Call(`_nlmixr2est_vaeOuterSolve_`, thVals, ebes, cols, cores, tol)
 }
 
 #' Build the nonparametric Psi (conditional-likelihood) matrix