+++ src/RcppExports.cpp
@@ -751,8 +751,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // vaeOuterSolve_
-RObject vaeOuterSolve_(NumericVector thVals, NumericMatrix ebes, List cols, int cores);
-RcppExport SEXP _nlmixr2est_vaeOuterSolve_(SEXP thValsSEXP, SEXP ebesSEXP, SEXP colsSEXP, SEXP coresSEXP) {
+RObject vaeOuterSolve_(NumericVector thVals, NumericMatrix ebes, List cols, int cores, double tol);
+RcppExport SEXP _nlmixr2est_vaeOuterSolve_(SEXP thValsSEXP, SEXP ebesSEXP, SEXP colsSEXP, SEXP coresSEXP, SEXP tolSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -760,7 +760,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< NumericMatrix >::type ebes(ebesSEXP);
     Rcpp::traits::input_parameter< List >::type cols(colsSEXP);
     Rcpp::traits::input_parameter< int >::type cores(coresSEXP);
-    rcpp_result_gen = Rcpp::wrap(vaeOuterSolve_(thVals, ebes, cols, cores));
+    Rcpp::traits::input_parameter< double >::type tol(tolSEXP);
+    rcpp_result_gen = Rcpp::wrap(vaeOuterSolve_(thVals, ebes, cols, cores, tol));
     return rcpp_result_gen;
 END_RCPP
 }